(PC-20703) feat(ThematicHome): refacto of all ThematicHeader in one file
