+++ src/features/home/components/headers/CategoryThematicHomeHeader.tsx
@@ -1,111 +1,68 @@
-import { useNavigation } from '@react-navigation/native'
-import React, { FunctionComponent, useCallback, useState } from 'react'
-import { Animated, Platform } from 'react-native'
-import styled, { useTheme } from 'styled-components/native'
+import React, { FunctionComponent } from 'react'
+import styled from 'styled-components/native'
 
+import { BlackGradient } from 'features/home/components/BlackGradient'
+import { HEADER_BLACK_BACKGROUND_HEIGHT } from 'features/home/components/constants'
+import { BlackBackground } from 'features/home/components/headers/BlackBackground'
 import { CategoryThematicHeader } from 'features/home/types'
-import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { homeNavConfig } from 'features/navigation/TabBar/helpers'
-import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
-import { RoundedButton } from 'ui/components/buttons/RoundedButton'
-import { BlurHeader } from 'ui/components/headers/BlurHeader'
-import { Spacer, Typo } from 'ui/theme'
-import { useCustomSafeInsets } from 'ui/theme/useCustomSafeInsets'
+import { getSpacing, Spacer, Typo } from 'ui/theme'
 
-type CategoryThematicHeaderProps = Omit<CategoryThematicHeader, 'type'> & {
-  headerTransition: Animated.AnimatedInterpolation
-}
+const HEADER_HEIGHT = getSpacing(44)
+
+type CategoryThematicHeaderProps = Omit<CategoryThematicHeader, 'type'>
 
 export const CategoryThematicHomeHeader: FunctionComponent<CategoryThematicHeaderProps> = ({
   title,
-  headerTransition,
+  subtitle,
+  imageUrl,
 }) => {
-  const { navigate } = useNavigation<UseNavigationType>()
-  const onGoBack = useCallback(() => navigate(...homeNavConfig), [navigate])
-  const { top } = useCustomSafeInsets()
-
-  const [ariaHiddenTitle, setAriaHiddenTitle] = useState(true)
-  headerTransition.addListener((opacity) => setAriaHiddenTitle(opacity.value !== 1))
-
-  const theme = useTheme()
-  const { animationState, styledContainer, blurContainerNative } = getAnimationState(
-    theme,
-    headerTransition
-  )
-
   return (
     <React.Fragment>
-      <HeaderContainer style={styledContainer} safeAreaTop={top}>
-        <Spacer.TopScreen />
-        <BlurNativeContainer style={blurContainerNative} safeAreaTop={top}>
-          <BlurHeader blurAmount={8} />
-        </BlurNativeContainer>
-        <Spacer.Column numberOfSpaces={2} />
-        <Row>
-          <Spacer.Row numberOfSpaces={6} />
-          <RoundedButton
-            animationState={animationState}
-            iconName="back"
-            onPress={onGoBack}
-            accessibilityLabel="Revenir en arrière"
-            finalColor={theme.colors.black}
-          />
-          <Spacer.Row numberOfSpaces={3} />
-          <Spacer.Row testID="leftShareIconPlaceholder" numberOfSpaces={10} />
-          <Spacer.Flex />
-          <Title
-            testID="offerHeaderName"
-            style={{ opacity: headerTransition }}
-            accessibilityHidden={ariaHiddenTitle}>
-            <Body>{title}</Body>
-          </Title>
-          <Spacer.Flex />
-          <Spacer.Row testID="rightSpacer" numberOfSpaces={25} />
-        </Row>
-        <Spacer.Column numberOfSpaces={2} />
-      </HeaderContainer>
+      <ImageBackground source={{ uri: imageUrl }}>
+        <TextContainer>
+          <BlackGradient height={HEADER_BLACK_BACKGROUND_HEIGHT} />
+          <BlackBackground>
+            <TitleContainer>
+              <Title numberOfLines={2}>{title}</Title>
+              {!!subtitle && (
+                <React.Fragment>
+                  <Subtitle numberOfLines={2}>{subtitle}</Subtitle>
+                  <Spacer.Column numberOfSpaces={1} />
+                </React.Fragment>
+              )}
+            </TitleContainer>
+          </BlackBackground>
+        </TextContainer>
+      </ImageBackground>
     </React.Fragment>
   )
 }
 
-const HeaderContainer = styled(Animated.View)<{ safeAreaTop: number }>(
-  ({ theme, safeAreaTop }) => ({
-    position: 'absolute',
-    top: 0,
-    height: theme.appBarHeight + safeAreaTop,
-    width: '100%',
-    zIndex: theme.zIndex.header,
-    borderBottomColor: theme.colors.greyLight,
-    borderBottomWidth: 1,
-  })
-)
-
-const BlurNativeContainer = styled(Animated.View)<{ safeAreaTop: number }>(
-  ({ theme, safeAreaTop }) => ({
-    position: 'absolute',
-    height: theme.appBarHeight + safeAreaTop,
-    top: 0,
-    left: 0,
-    right: 0,
-    overflow: 'hidden',
-  })
-)
-
-const Row = styled.View({
-  flex: 1,
-  flexDirection: 'row',
-  alignItems: 'center',
+const ImageBackground = styled.ImageBackground({
+  height: HEADER_HEIGHT,
+  marginBottom: getSpacing(6),
 })
 
-const Title = styled(Animated.Text).attrs({
-  numberOfLines: 2,
-})(({ theme }) => ({
-  flexShrink: 1,
-  textAlign: 'center',
+const TextContainer = styled.View({
+  position: 'absolute',
+  bottom: 0,
+  left: 0,
+  right: 0,
+})
+
+const Subtitle = styled(Typo.Title4)(({ theme }) => ({
   color: theme.colors.white,
-  ...(Platform.OS === 'web' ? { whiteSpace: 'pre-wrap' } : {}),
 }))
 
-const Body = styled(Typo.Body)(({ theme }) => ({
-  color: theme.colors.black,
+const Title = styled(Typo.Title1)(({ theme }) => ({
+  color: theme.colors.white,
+}))
+
+const TitleContainer = styled.View(({ theme }) => ({
+  // The size of the title block should not exceed two lines of title and one of subtitle
+  maxHeight:
+    parseInt(theme.typography.title1.lineHeight) * 2 +
+    getSpacing(1) +
+    parseInt(theme.typography.title4.lineHeight),
+  overflow: 'hidden',
 }))
