+++ src/features/home/components/headers/DefaultThematicHomeHeader.native.test.tsx
@@ -1,24 +1,12 @@
 import React from 'react'
-import { Animated } from 'react-native'
 
-import { navigate } from '__mocks__/@react-navigation/native'
 import { DefaultThematicHomeHeader } from 'features/home/components/headers/DefaultThematicHomeHeader'
-import { navigateToHomeConfig } from 'features/navigation/helpers'
-import { fireEvent, render, screen } from 'tests/utils'
-
-const animatedValue = new Animated.Value(0)
-const HeaderInterpolation = animatedValue.interpolate({
-  inputRange: [0, 1],
-  outputRange: [0, 1],
-})
+import { render, screen } from 'tests/utils'
 
 describe('DefaultThematicHomeHeader', () => {
-  it('should navigate to home page on press go back button', () => {
-    render(<DefaultThematicHomeHeader headerTitle="toto" headerTransition={HeaderInterpolation} />)
-    const backButton = screen.getByTestId('Revenir en arrière')
-
-    fireEvent.press(backButton)
-
-    expect(navigate).toBeCalledWith(navigateToHomeConfig.screen, navigateToHomeConfig.params)
+  it('should show title and sub title if render correctly', async () => {
+    render(<DefaultThematicHomeHeader headerSubtitle={'Un sous-titre'} headerTitle={'Un titre'} />)
+    expect(await screen.findByText('Un titre')).toBeTruthy()
+    expect(screen.getByText('Un sous-titre')).toBeTruthy()
   })
 })
