+++ src/features/home/components/headers/DefaultThematicHomeHeader.tsx
@@ -1,110 +1,49 @@
-import { useNavigation } from '@react-navigation/native'
-import React, { FunctionComponent, useCallback, useState } from 'react'
-import { Animated, Platform } from 'react-native'
-import styled, { useTheme } from 'styled-components/native'
+import React, { FunctionComponent } from 'react'
+import { Platform } from 'react-native'
+import styled from 'styled-components/native'
 
-import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { homeNavConfig } from 'features/navigation/TabBar/helpers'
-import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
-import { RoundedButton } from 'ui/components/buttons/RoundedButton'
-import { BlurHeader } from 'ui/components/headers/BlurHeader'
-import { Spacer, Typo } from 'ui/theme'
-import { useCustomSafeInsets } from 'ui/theme/useCustomSafeInsets'
+import { getSpacing, Spacer, Typo } from 'ui/theme'
 
 export interface DefaultThematicHomeHeaderProps {
-  headerTransition: Animated.AnimatedInterpolation
   headerTitle?: string
+  headerSubtitle?: string
 }
 export const DefaultThematicHomeHeader: FunctionComponent<DefaultThematicHomeHeaderProps> = ({
-  headerTransition,
   headerTitle,
+  headerSubtitle,
 }) => {
-  const { navigate } = useNavigation<UseNavigationType>()
-  const onGoBack = useCallback(() => navigate(...homeNavConfig), [navigate])
-  const { top } = useCustomSafeInsets()
-
-  const [ariaHiddenTitle, setAriaHiddenTitle] = useState(true)
-  headerTransition.addListener((opacity) => setAriaHiddenTitle(opacity.value !== 1))
-
-  const theme = useTheme()
-  const { animationState, styledContainer, blurContainerNative } = getAnimationState(
-    theme,
-    headerTransition
-  )
-
   return (
-    <React.Fragment>
-      <HeaderContainer style={styledContainer} safeAreaTop={top}>
-        <Spacer.TopScreen />
-        <BlurNativeContainer style={blurContainerNative} safeAreaTop={top}>
-          <BlurHeader blurAmount={8} />
-        </BlurNativeContainer>
-        <Spacer.Column numberOfSpaces={2} />
-        <Row>
-          <Spacer.Row numberOfSpaces={6} />
-          <RoundedButton
-            animationState={animationState}
-            iconName="back"
-            onPress={onGoBack}
-            accessibilityLabel="Revenir en arrière"
-            finalColor={theme.colors.black}
-          />
-          <Spacer.Row numberOfSpaces={3} />
-          <Spacer.Row testID="leftShareIconPlaceholder" numberOfSpaces={10} />
-          <Spacer.Flex />
-          <Title
-            testID="offerHeaderName"
-            style={{ opacity: headerTransition }}
-            accessibilityHidden={ariaHiddenTitle}>
-            <Body>{headerTitle}</Body>
-          </Title>
-          <Spacer.Flex />
-          <Spacer.Row testID="rightSpacer" numberOfSpaces={25} />
-        </Row>
-        <Spacer.Column numberOfSpaces={2} />
-      </HeaderContainer>
-    </React.Fragment>
+    <Container>
+      {headerTitle ? (
+        <React.Fragment>
+          <Spacer.Column numberOfSpaces={4} />
+          <TitleContainer>
+            <Typo.Title1 numberOfLines={2}>{headerTitle}</Typo.Title1>
+            {headerSubtitle ? (
+              <React.Fragment>
+                <Spacer.Column numberOfSpaces={2} />
+                <Typo.Body numberOfLines={2}>{headerSubtitle}</Typo.Body>
+              </React.Fragment>
+            ) : null}
+          </TitleContainer>
+          <Spacer.Column numberOfSpaces={6} />
+        </React.Fragment>
+      ) : null}
+    </Container>
   )
 }
 
-const HeaderContainer = styled(Animated.View)<{ safeAreaTop: number }>(
-  ({ theme, safeAreaTop }) => ({
-    position: 'absolute',
-    top: 0,
-    height: theme.appBarHeight + safeAreaTop,
-    width: '100%',
-    zIndex: theme.zIndex.header,
-    borderBottomColor: theme.colors.greyLight,
-    borderBottomWidth: 1,
-  })
-)
-
-const BlurNativeContainer = styled(Animated.View)<{ safeAreaTop: number }>(
-  ({ theme, safeAreaTop }) => ({
-    position: 'absolute',
-    height: theme.appBarHeight + safeAreaTop,
-    top: 0,
-    left: 0,
-    right: 0,
-    overflow: 'hidden',
-  })
-)
-
-const Row = styled.View({
-  flex: 1,
-  flexDirection: 'row',
-  alignItems: 'center',
-})
-
-const Title = styled(Animated.Text).attrs({
-  numberOfLines: 2,
-})(({ theme }) => ({
-  flexShrink: 1,
-  textAlign: 'center',
-  color: theme.colors.white,
-  ...(Platform.OS === 'web' ? { whiteSpace: 'pre-wrap' } : {}),
+const Container = styled.View(({ theme }) => ({
+  marginTop: Platform.OS == 'web' ? getSpacing(6) : getSpacing(2),
+  marginHorizontal: theme.contentPage.marginHorizontal,
+  marginBottom: getSpacing(2),
 }))
 
-const Body = styled(Typo.Body)(({ theme }) => ({
-  color: theme.colors.black,
+const TitleContainer = styled.View(({ theme }) => ({
+  // The size of the title block should not exceed two lines of title and one of subtitle
+  maxHeight:
+    parseInt(theme.typography.title1.lineHeight) * 2 +
+    getSpacing(2) +
+    parseInt(theme.typography.body.lineHeight),
+  overflow: 'hidden',
 }))
