+++ src/features/home/components/headers/HighlightThematicHomeHeader.tsx
@@ -1,112 +1,114 @@
-import { useNavigation } from '@react-navigation/native'
-import React, { FunctionComponent, useCallback, useState } from 'react'
-import { Animated, Platform } from 'react-native'
-import styled, { useTheme } from 'styled-components/native'
+import React, { FunctionComponent } from 'react'
+import styled from 'styled-components/native'
 
+import { BlackGradient } from 'features/home/components/BlackGradient'
+import { HEADER_BLACK_BACKGROUND_HEIGHT } from 'features/home/components/constants'
+import { BlackBackground } from 'features/home/components/headers/BlackBackground'
+import { computeDateRangeDisplay } from 'features/home/components/helpers/computeDateRangeDisplay'
 import { HighlightThematicHeader } from 'features/home/types'
-import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { homeNavConfig } from 'features/navigation/TabBar/helpers'
-import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
-import { RoundedButton } from 'ui/components/buttons/RoundedButton'
-import { BlurHeader } from 'ui/components/headers/BlurHeader'
-import { Spacer, Typo } from 'ui/theme'
+import { getSpacing, Spacer, Typo } from 'ui/theme'
 import { useCustomSafeInsets } from 'ui/theme/useCustomSafeInsets'
 
-type HighligthThematicHeaderProps = Omit<HighlightThematicHeader, 'type'> & {
-  headerTransition: Animated.AnimatedInterpolation
+type HighligthThematicHeaderProps = Omit<HighlightThematicHeader, 'type'>
+
+type IntroductionProps = {
+  title: string
+  paragraph: string
 }
+
+const DESKTOP_HEADER_HEIGHT = getSpacing(100)
+const MOBILE_HEADER_HEIGHT = getSpacing(70)
 
 export const HighlightThematicHomeHeader: FunctionComponent<HighligthThematicHeaderProps> = ({
   title,
-  headerTransition,
+  subtitle,
+  imageUrl,
+  beginningDate,
+  endingDate,
+  introductionTitle,
+  introductionParagraph,
 }) => {
-  const { navigate } = useNavigation<UseNavigationType>()
-  const onGoBack = useCallback(() => navigate(...homeNavConfig), [navigate])
   const { top } = useCustomSafeInsets()
 
-  const [ariaHiddenTitle, setAriaHiddenTitle] = useState(true)
-  headerTransition.addListener((opacity) => setAriaHiddenTitle(opacity.value !== 1))
+  const dateRange = computeDateRangeDisplay(beginningDate, endingDate)
 
-  const theme = useTheme()
-  const { animationState, styledContainer, blurContainerNative } = getAnimationState(
-    theme,
-    headerTransition
-  )
+  const shouldShowIntroduction = !!introductionTitle && !!introductionParagraph
 
   return (
     <React.Fragment>
-      <HeaderContainer style={styledContainer} safeAreaTop={top}>
-        <Spacer.TopScreen />
-        <BlurNativeContainer style={blurContainerNative} safeAreaTop={top}>
-          <BlurHeader blurAmount={8} />
-        </BlurNativeContainer>
-        <Spacer.Column numberOfSpaces={2} />
-        <Row>
-          <Spacer.Row numberOfSpaces={6} />
-          <RoundedButton
-            animationState={animationState}
-            iconName="back"
-            onPress={onGoBack}
-            accessibilityLabel="Revenir en arrière"
-            finalColor={theme.colors.black}
-          />
-          <Spacer.Row numberOfSpaces={3} />
-          <Spacer.Row testID="leftShareIconPlaceholder" numberOfSpaces={10} />
-          <Spacer.Flex />
-          <Title
-            testID="offerHeaderName"
-            style={{ opacity: headerTransition }}
-            accessibilityHidden={ariaHiddenTitle}>
-            <Body>{title}</Body>
-          </Title>
-          <Spacer.Flex />
-          <Spacer.Row testID="rightSpacer" numberOfSpaces={25} />
-        </Row>
-        <Spacer.Column numberOfSpaces={2} />
-      </HeaderContainer>
+      <ImageBackground source={{ uri: imageUrl }}>
+        <DateRangeCaptionContainer statusBarHeight={top}>
+          <DateRangeCaption>{dateRange}</DateRangeCaption>
+        </DateRangeCaptionContainer>
+        <TextContainer>
+          <BlackGradient height={HEADER_BLACK_BACKGROUND_HEIGHT} />
+          <BlackBackground>
+            {!!subtitle && (
+              <React.Fragment>
+                <Subtitle numberOfLines={1}>{subtitle}</Subtitle>
+                <Spacer.Column numberOfSpaces={1} />
+              </React.Fragment>
+            )}
+            <Title numberOfLines={2}>{title}</Title>
+          </BlackBackground>
+        </TextContainer>
+      </ImageBackground>
+      {shouldShowIntroduction ? (
+        <Introduction title={introductionTitle} paragraph={introductionParagraph} />
+      ) : null}
     </React.Fragment>
   )
 }
 
-const HeaderContainer = styled(Animated.View)<{ safeAreaTop: number }>(
-  ({ theme, safeAreaTop }) => ({
-    position: 'absolute',
-    top: 0,
-    height: theme.appBarHeight + safeAreaTop,
-    width: '100%',
-    zIndex: theme.zIndex.header,
-    borderBottomColor: theme.colors.greyLight,
-    borderBottomWidth: 1,
-    backgroundColor: 'red',
-  })
+const Introduction = ({ title, paragraph }: IntroductionProps) => (
+  <React.Fragment>
+    <IntroductionContainer>
+      <Typo.Title4 numberOfLines={3}>{title}</Typo.Title4>
+      <Spacer.Column numberOfSpaces={4} />
+      <Typo.Body>{paragraph}</Typo.Body>
+    </IntroductionContainer>
+    <Spacer.Column numberOfSpaces={6} />
+    <Divider />
+    <Spacer.Column numberOfSpaces={6} />
+  </React.Fragment>
 )
 
-const BlurNativeContainer = styled(Animated.View)<{ safeAreaTop: number }>(
-  ({ theme, safeAreaTop }) => ({
-    position: 'absolute',
-    height: theme.appBarHeight + safeAreaTop,
-    top: 0,
-    left: 0,
-    right: 0,
-    overflow: 'hidden',
-  })
-)
-
-const Row = styled.View({
-  flex: 1,
-  flexDirection: 'row',
-  alignItems: 'center',
-})
-
-const Title = styled(Animated.Text).attrs({
-  numberOfLines: 2,
-})(({ theme }) => ({
-  flexShrink: 1,
-  textAlign: 'center',
-  color: theme.colors.white,
-  ...(Platform.OS === 'web' ? { whiteSpace: 'pre-wrap' } : {}),
+const ImageBackground = styled.ImageBackground(({ theme }) => ({
+  height: theme.isDesktopViewport ? DESKTOP_HEADER_HEIGHT : MOBILE_HEADER_HEIGHT,
+  marginBottom: getSpacing(6),
 }))
 
-const Body = styled(Typo.Body)(({ theme }) => ({
-  color: theme.colors.black,
+const DateRangeCaptionContainer = styled.View<{ statusBarHeight: number }>(
+  ({ theme, statusBarHeight }) => ({
+    backgroundColor: theme.colors.black,
+    position: 'absolute',
+    top: statusBarHeight + getSpacing(6),
+    right: getSpacing(6),
+    borderRadius: getSpacing(2),
+    paddingVertical: getSpacing(1),
+    paddingHorizontal: getSpacing(2),
+  })
+)
+
+const DateRangeCaption = styled(Typo.Caption)(({ theme }) => ({
+  color: theme.colors.white,
+}))
+
+const TextContainer = styled.View({ position: 'absolute', bottom: 0, left: 0, right: 0 })
+
+const Subtitle = styled(Typo.Title4)(({ theme }) => ({
+  color: theme.colors.white,
+}))
+
+const Title = styled(Typo.Title1)(({ theme }) => ({
+  color: theme.colors.white,
+}))
+
+const IntroductionContainer = styled.View({
+  paddingHorizontal: getSpacing(6),
+})
+
+const Divider = styled.View(({ theme }) => ({
+  height: getSpacing(1),
+  backgroundColor: theme.colors.greyLight,
 }))
