+++ src/features/internal/cheatcodes/pages/ThematicHomeHeaderCheatcode/CategoryThematicHomeHeaderCheatcode.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react'
 import styled from 'styled-components/native'
 
 import { CategoryThematicHomeHeader } from 'features/home/components/headers/CategoryThematicHomeHeader'
-import { CategoryThematicHomeSubHeader } from 'features/home/components/headers/CategoryThematicHomeSubHeader'
+import { ThematicHomeHeader } from 'features/home/components/headers/ThematicHomeHeader'
 import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
 
 export const CategoryThematicHomeHeaderCheatcode: FunctionComponent = () => {
@@ -10,13 +10,8 @@ export const CategoryThematicHomeHeaderCheatcode: FunctionComponent = () => {
 
   return (
     <Container>
+      <ThematicHomeHeader headerTransition={headerTransition} title={'un titre'} />
       <CategoryThematicHomeHeader
-        imageUrl=""
-        subtitle=""
-        title={'Un titre'}
-        headerTransition={headerTransition}
-      />
-      <CategoryThematicHomeSubHeader
         imageUrl="https://images.ctfassets.net/2bg01iqy0isv/5PmtxKY77rq0nYpkCFCbrg/4daa8767efa35827f22bb86e5fc65094/photo-lion_noir-et-blanc_laurent-breillat-610x610.jpeg"
         subtitle={'Un sous-titre'}
         title={'Un titre'}
